Cryptographically Signed tokens
Uses CMS to create tokens that can be verified without network calls.

Tokens encapsulate authorization information.
This includes user name and roles in JSON.
The JSON document info is cryptographically signed with a private key
from Keystone, in accordance with the Cryptographic Message Syntax (CMS)
in DER format and then Base64 encoded.  The header, footer, and line breaks
are stripped to minimize the size,  and slashes which are  invalid in Base64
are converted to hyphens.

Since signed tokens are not validated against the Keystone server,  they
continue to be valid until the expiration time.  This means that even if a user
has their roles revoked or their account disabled, those changes will not take
effect until their token times out.  The prototype for this is Kerberos, which
has the same limitation, and has funtioned sucessfully with it for decades.  It
is possible to set the token time out for much shorter than the default of 8
hours, but that may mean that users tokens will time out prior to completion
of long running tasks.

This should be a drop in replacement for the current token production code.
Although  the signed token is longer than the older format, the token is still
a unique stream of Alpha-Numeric characters.

The auth token middle_ware is capable of handling both uuid and signed tokens.

To start with, the PKI functionality is disabled.  This will keep from breaking
the existing deployments.  However,  it can be enabled with the config value:

[signing]
disable_pki = False

The 'id_hash' column is added to the SQL schema because SQL alchemy insists on
each table having a primary key.  However primary keys are limited to roughly
250 Characters (768 Bytes,  but there is more than 1 varchar per byte) so the
ID field cannot be used as the primary key anymore.  id_hash is a hash of the
id column, and should be used for lookups as it is indexed.

middleware/auth_token.py needs to stand alone in the other services, and uses
keystone.common.cms in order to verify tokens.
Token needs to have all of the data from the original authenticate code
contained in the signed document, as the authenticate RPC will no longer
be called in mand cases.

The datetime of expiry is signed in the token.

The certificates are accessible via web APIs.  On the remote service side,
certificates needed to authenitcate tokens are stored in /tmp/keystone-signing
by default.  Remote systems use Paste API to read configuration values.
Certificates are retrieved only if they are not on the local system.

When authenticating in Keystone systems, it still does the Database checks for
token presence.  This allows Keystone to continue to enforce Timeout and
disabled users.

The service catalog has been added to the  signed token.  Although this greatly
increases the size of the token,  it makes it consistant with what is fetched
during the token authenticate checks

This change also fixes time variations in expiry test.  Although unrelated to
the above changes, it was making testing very frustrating.

For the database Upgrade scripts, we now only  bring 'token' up to V1 in 001
script.  This makes it possible to use the same 002 script for both upgrade
and initializing a new database.

Upon upgrade, the current UUID tokens are retained in the id_hash and id fields.
The mechanisms to verify uuid tokens work the same as before.  On downgrade,
token_ids are dropped.

Takes into account changes for "Raise unauthorized if tenant disabled"

    Bug 1003962

+++ keystone/token/backends/sql.py
@@ -15,8 +15,11 @@
 # under the License.
 
 import copy
+import datetime
+import hashlib
+import uuid
 
-from keystone.common import sql
+from keystone.common import sql, cms
 from keystone import exception
 from keystone.openstack.common import timeutils
 from keystone import token
@@ -24,7 +27,8 @@ from keystone import token
 
 class TokenModel(sql.ModelBase, sql.DictBase):
     __tablename__ = 'token'
-    id = sql.Column(sql.String(64), primary_key=True)
+    id_hash = sql.Column(sql.String(64), primary_key=True)
+    id = sql.Column(sql.String(1024))
     expires = sql.Column(sql.DateTime(), default=None)
     extra = sql.Column(sql.JsonBlob())
 
@@ -33,13 +37,14 @@ class TokenModel(sql.ModelBase, sql.DictBase):
         # shove any non-indexed properties into extra
         extra = copy.deepcopy(token_dict)
         data = {}
-        for k in ('id', 'expires'):
+        for k in ('id_hash', 'id', 'expires'):
             data[k] = extra.pop(k, None)
         data['extra'] = extra
         return cls(**data)
 
     def to_dict(self):
         out = copy.deepcopy(self.extra)
+        out['id_hash'] = self.id
         out['id'] = self.id
         out['expires'] = self.expires
         return out
@@ -49,12 +54,21 @@ class Token(sql.Base, token.Driver):
     # Public interface
     def get_token(self, token_id):
         session = self.get_session()
-        token_ref = session.query(TokenModel).filter_by(id=token_id).first()
-        now = timeutils.utcnow()
+        token_ref = session.query(TokenModel)\
+            .filter_by(id_hash=self.token_to_key(token_id)).first()
+        now = datetime.datetime.utcnow()
         if token_ref and (not token_ref.expires or now < token_ref.expires):
             return token_ref.to_dict()
         else:
             raise exception.TokenNotFound(token_id=token_id)
+
+    def token_to_key(self, token_id):
+        if len(token_id) > cms.UUID_TOKEN_LENGTH:
+            hash = hashlib.md5()
+            hash.update(token_id)
+            return hash.hexdigest()
+        else:
+            return token_id
 
     def create_token(self, token_id, data):
         data_copy = copy.deepcopy(data)
@@ -62,8 +76,7 @@ class Token(sql.Base, token.Driver):
             data_copy['expires'] = self._get_default_expire_time()
 
         token_ref = TokenModel.from_dict(data_copy)
-        token_ref.id = token_id
-
+        token_ref.id_hash = self.token_to_key(token_id)
         session = self.get_session()
         with session.begin():
             session.add(token_ref)
@@ -72,6 +85,12 @@ class Token(sql.Base, token.Driver):
 
     def delete_token(self, token_id):
         session = self.get_session()
+        token_ref = session.query(TokenModel)\
+                           .filter_by(id_hash=self.token_to_key(token_id))\
+                           .first()
+        if not token_ref:
+            raise exception.TokenNotFound(token_id=token_id)
+
         with session.begin():
             if not session.query(TokenModel).filter_by(id=token_id).delete():
                 raise exception.TokenNotFound(token_id=token_id)
